pyproject.toml
[project]
name = "song-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lyricsgenius>=3.7.5",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
]
